requirements.txt
imgaug==0.2.8
progressbar2==3.39.3
pandas==0.24.2
more-itertools==7.0.0
fire==0.1.3
pyjokes==0.5.0
wget==3.2
checksumdir==1.1.6
bs4==0.0.1
requests==2.22.0
plotly==3.9.0